loss type form float to half

+++ tests/test_gemini/test_mem_tracer_paramOP.py
@@ -12,7 +12,7 @@ def run_fwd_bwd(model, data, label, criterion, enable_autocast=False):
             loss = criterion(y, label)
         else:
             loss = model(data, label)
-        loss = loss.float()
+        loss = loss.half()
     model.backward(loss)
 
 def run_param_wrapper_testing():
